thêm 1 vài thông tin

+++ src/main/java/common/models/Bidder.java
@@ -3,7 +3,7 @@ package common.models;
 import java.util.ArrayList;
 
 public class Bidder extends User {
-    private int balance = 0;
+    private double balance = 0;
     private ArrayList<String> bidHistory = new ArrayList<>();
 
     public Bidder(int id, String loginUsername, double balance) {
